# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Edge Solutions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Keyword {
  id: string;
  keyword: string;
  volume: number;
  difficulty: string;
  intent: string;
  service: string;
  city: string;
  state: string;
}

export type KeywordsBulkInputItemsItem = {
  keyword: string;
  volume: number;
  difficulty: string;
  intent: string;
  service: string;
  city: string;
  state: string;
};

export interface KeywordsBulkInput {
  items: KeywordsBulkInputItemsItem[];
}

export interface ArticleDraft {
  id: string;
  title: string;
  keyword: string;
  /** @nullable */
  keywordId?: string | null;
  service: string;
  project: string;
  body: string;
  metaTitle: string;
  metaDescription: string;
  slug: string;
  status: string;
  /** @nullable */
  scheduledFor?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  publishedUrl?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  /** @nullable */
  verifiedLiveAt?: string | null;
  /** @nullable */
  lastStatusCode?: number | null;
  /** @nullable */
  lastCheckedAt?: string | null;
}

export interface ArticlesBulkInput {
  items: ArticleDraft[];
}

export interface ArticleUpdate {
  title?: string;
  body?: string;
  metaTitle?: string;
  metaDescription?: string;
  slug?: string;
  status?: string;
  /** @nullable */
  scheduledFor?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  publishedUrl?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  /** @nullable */
  verifiedLiveAt?: string | null;
  /** @nullable */
  lastStatusCode?: number | null;
  /** @nullable */
  lastCheckedAt?: string | null;
}

export interface ArticleAsset {
  id: string;
  articleId: string;
  channel: string;
  body: string;
  status: string;
  /** @nullable */
  publishedUrl?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  updatedAt?: string;
}

export type ArticleAssetsBulkInputAssetsItem = {
  channel: string;
  body: string;
  status?: string;
  /** @nullable */
  errorMessage?: string | null;
};

export interface ArticleAssetsBulkInput {
  assets: ArticleAssetsBulkInputAssetsItem[];
}

export interface ArticleAssetUpdate {
  body?: string;
  status?: string;
  /** @nullable */
  publishedUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface KeywordGenInput {
  businessName: string;
  industry: string;
  city: string;
  state: string;
  mainServices: string;
  targetCustomers?: string;
}

export interface KeywordGenResult {
  keywords: Keyword[];
}

export interface ArticleGenInput {
  title: string;
  keyword: string;
  service: string;
  businessName: string;
  industry: string;
  city: string;
  state: string;
  mainServices?: string;
  targetCustomers?: string;
}

export interface ArticleGenResult {
  body: string;
}

export interface ArticleAssetsGenInput {
  title: string;
  keyword: string;
  service?: string;
  businessName?: string;
  city?: string;
  state?: string;
  body?: string;
}

export type ArticleAssetsGenResultAssetsItem = {
  channel: string;
  body: string;
  status: string;
  /** @nullable */
  errorMessage?: string | null;
};

export interface ArticleAssetsGenResult {
  assets: ArticleAssetsGenResultAssetsItem[];
}

export interface RepurposeInput {
  businessName: string;
  service: string;
  city: string;
  state?: string;
  keyword: string;
}

export type RepurposeResultAssetsItem = {
  channel: string;
  label: string;
  body: string;
};

export interface RepurposeResult {
  assets: RepurposeResultAssetsItem[];
}