# lib/api-zod/src/generated/types/articleAssetsBulkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Edge Solutions API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleAssetsBulkInputAssetsItem } from './articleAssetsBulkInputAssetsItem';

export interface ArticleAssetsBulkInput {
  assets: ArticleAssetsBulkInputAssetsItem[];
}