# lib/api-zod/src/generated/types/articleAssetsGenInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Edge Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleAssetsGenInput {
  title: string;
  keyword: string;
  service?: string;
  businessName?: string;
  city?: string;
  state?: string;
  body?: string;
}