# lib/api-zod/src/generated/types/articleAssetsGenResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Edge Solutions API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleAssetsGenResultAssetsItem } from './articleAssetsGenResultAssetsItem';

export interface ArticleAssetsGenResult {
  assets: ArticleAssetsGenResultAssetsItem[];
}